# openapi.yaml
openapi: 3.1.0
info:
  title: AndClaw API
  description: Personal AI Agent — Command Center backend API
  version: 2.0.0
  contact:
    name: AndClaw
servers:
  - url: https://andclaw.onrender.com
    description: Production (Render)
  - url: http://localhost:3000
    description: Local Development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '401':
          description: Invalid password
  /api/auth/bootstrap:
    post:
      tags: [Auth]
      summary: First-time system setup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:    { type: string, minLength: 8 }
                tokenSecret: { type: string, minLength: 24 }
      responses:
        '200':
          description: Bootstrap successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:       { type: string }
                  tokenSecret: { type: string }
        '409':
          description: Already bootstrapped
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Verify token and get current user
      responses:
        '200':
          description: Current user payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  id: { type: string }
                  email: { type: [string, 'null'] }
                  fullName: { type: [string, 'null'] }
                  name: { type: [string, 'null'] }
                  company: { type: [string, 'null'] }
                  role: { type: [string, 'null'] }
                  photoUrl: { type: [string, 'null'] }
                  profile:
                    type: object
                    additionalProperties: { type: string }
        '401':
          description: Unauthorized
...
  /api/memory:
    post:
      tags: [Memory]
      summary: Create a memory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, content]
              properties:
                type: { type: string }
                content: { type: string }
                source_type: { type: string }
                source_id: { type: string }
      responses:
        '201':
          description: Memory item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  id: { type: integer }
                  item:
                    type: object
                    additionalProperties: true
  /api/knowledge:
    post:
      tags: [Memory]
      summary: Create a knowledge item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string }
                title: { type: string, description: "Alias for content" }
                content: { type: string, description: "Main text content" }
                source_type: { type: string }
                source_id: { type: string }
      responses:
        '201':
          description: Knowledge item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  id: { type: integer }
                  item:
                    type: object
                    additionalProperties: true
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get user profile
      responses:
        '200':
          description: Profile data

    post:
      tags: [Profile]
      summary: Update user profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: { type: string }
                email:    { type: string, format: email }
                company:  { type: string }
                role:     { type: string }
                photoUrl: { type: string, format: uri }
      responses:
        '200':
          description: Profile updated
  /api/preferences:
    get:
      tags: [Profile]
      summary: Get user preferences
      responses:
        '200':
          description: Preferences

    post:
      tags: [Profile]
      summary: Update preferences
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:           { type: string, enum: [dark, light, auto] }
                language:        { type: string }
                notifyEmail:     { type: string }
                notifyPush:      { type: string }
      responses:
        '200':
          description: Preferences updated

  # ── PUSH ────────────────────────────────────────────────────
...